LPD-27899 Rename, use camel case format
