+++ modules/test/playwright/tests/blogs-web/utils/blogsCategorizedFriendlyUrlSetup.ts
@@ -4,7 +4,7 @@
  */
 
 import {createAssetPublisherAndConfigure} from './createAssetPublisherAndConfigure';
-import {TCategory, createCategories} from '../../../helpers/createCategories';
+import {TCategory, createCategories} from '../../../helpers/CreateCategories';
 import {createDPTandMarkAsDefault} from './createDPTandMarkAsDefault';
 
 import type {ApiHelpers} from '../../../helpers/ApiHelpers';
